requirements.txt
Flask
Flask-Cors
torch
torchvision
Pillow
gunicorn